webstorage/iframe/local_security_iframe.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_throws("SecurityError", function() {
    parent.localStorage.clear();
  });
});
</script>
